# api/swagger.yaml
openapi: 3.0.0
info:
  title: Simple Task API
  description: This is a Task API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/anshul2/crowdstrike/1.0.0
  description: SwaggerHub API Auto Mocking
- url: http://localhost:8080/crowdstrike/v1
  description: SwaggerHub API Auto Mocking
tags:
- name: Tasks
  description: Operations available to regular developers
- name: File Search Tasks API
  description: Secured Admin-only calls
paths:
  /tasks/{taskId}:
    get:
      tags:
      - Tasks
      summary: get task details by ID
      description: |
        By passing in the appropriate options, you can search for
        available inventory in the system
      operationId: searchTasks
      parameters:
      - name: task Id to be searched
        in: path
        description: Task
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TasksResponse'
        "400":
          description: bad input parameter
        "404":
          description: bad input parameter
  /tasks:
    post:
      tags:
      - Tasks
      summary: adds an task item
      description: Adds a new File processing task
      operationId: addTask
      requestBody:
        description: Task item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTasksRequest'
      responses:
        "202":
          description: item created
        "400":
          description: "invalid input, object invalid"
        "409":
          description: an existing item already exists
  /search/filesByIP:
    get:
      tags:
      - File Search Tasks API
      summary: get task details by ID
      description: "By passing in the appropriate options, you can search for\navailable\
        \ files in the system which has same IP \n"
      operationId: searchFilesByIP
      parameters:
      - name: ip
        in: query
        description: Mandatory to lookup files by this IP. LAter impletement pagination
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchFilesReponse'
        "400":
          description: bad input parameter
components:
  schemas:
    CreateTasksRequest:
      required:
      - fileID
      type: object
      properties:
        fileID:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
    TasksResponse:
      required:
      - fileID
      - taskCreationDate
      - taskID
      - taskResult
      - taskStatus
      type: object
      properties:
        taskID:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        taskCreationDate:
          type: integer
          format: int64
          example: 1656329451
        fileID:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        taskStatus:
          type: string
          enum:
          - in-progress
          - completed
        taskResult:
          type: array
          items:
            type: string
      example:
        taskResult:
        - taskResult
        - taskResult
        taskID: d290f1ee-6c54-4b01-90e6-d701748f0851
        taskStatus: in-progress
        taskCreationDate: 1656329451
        fileID: d290f1ee-6c54-4b01-90e6-d701748f0851
    SearchFilesReponse:
      required:
      - files
      type: object
      properties:
        files:
          type: array
          items:
            type: string
      example:
        files:
        - files
        - files