# swagger/model_create_tasks_request.go
/*
 * Simple Task API
 *
 * This is a Task API
 *
 * API version: 1.0.0
 * Contact: you@your-company.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type CreateTasksRequest struct {

	FileID string `json:"fileID"`
}